split menu repository interface
