+++ src/main/java/com/example/menu/repository/MenuQueryRepository.java
@@ -1,0 +1,9 @@
+package com.example.menu.repository;
+
+import com.example.menu.entity.Menu;
+
+import java.util.List;
+
+interface MenuQueryRepository {
+    List<Menu> findAllByShopId(int id);
+}
